Move local mediaPane code so that we have media3p and local at the same level.
