+++ assets/src/edit-story/components/library/panes/media/common/mediaElement.js
@@ -27,10 +27,10 @@ import { useFeature } from 'flagged';
 /**
  * Internal dependencies
  */
-import { useDropTargets } from '../../../../app';
-import getThumbnailUrl from '../../../../app/media/utils/getThumbnailUrl';
-import DropDownMenu from './dropDownMenu';
-import { ProviderType } from './providerType';
+import { useDropTargets } from '../../../../../app';
+import getThumbnailUrl from '../../../../../app/media/utils/getThumbnailUrl';
+import DropDownMenu from '../local/dropDownMenu';
+import { ProviderType } from '../media3p/providerType';
 
 const styledTiles = css`
   width: 100%;
