+++ assets/src/edit-story/components/library/panes/media/common/stories/mediaGallery.js
@@ -24,7 +24,7 @@ import styled from 'styled-components';
  * Internal dependencies
  */
 import MediaGallery from '../mediaGallery';
-import { ProviderType } from '../providerType';
+import { ProviderType } from '../../media3p/providerType';
 
 export default {
   title: 'Stories Editor/Components/Media Gallery',
