+++ assets/src/edit-story/components/library/panes/media/common/stories/test-image.jpg
