+++ assets/src/edit-story/components/library/panes/media/common/stories/test-poster.png
