+++ assets/src/edit-story/components/library/panes/media/local/index.js
