+++ assets/src/edit-story/components/library/test/deleteDialog.js
@@ -22,7 +22,7 @@ import { waitFor, fireEvent } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import DeleteDialog from '../panes/media/deleteDialog';
+import DeleteDialog from '../panes/media/local/deleteDialog';
 import { renderWithTheme } from '../../../testUtils';
 import ApiContext from '../../../app/api/context';
 import MediaContext from '../../../app/media/context';
